+++ js/ui/html_editor/modules/toolbar.js
@@ -20,8 +20,7 @@ import { titleize, camelize } from '../../../core/utils/inflector';
 import eventsEngine from '../../../events/core/events_engine';
 import { addNamespace } from '../../../events/utils/index';
 
-import { showCellPropertiesForm, showTablePropertiesForm } from '../ui/tableForms';
-import { getTableFormats, getTableOperationHandler, TABLE_OPERATIONS } from '../utils/table_helper';
+import { getTableFormats, TABLE_OPERATIONS, getFormatHandlers } from '../utils/table_helper';
 
 let ToolbarModule = BaseModule;
 
@@ -37,24 +36,22 @@ if(Quill) {
 
     const SELECTION_CHANGE_EVENT = 'selection-change';
 
-    const DIALOG_COLOR_CAPTION = 'dxHtmlEditor-dialogColorCaption';
-    const DIALOG_BACKGROUND_CAPTION = 'dxHtmlEditor-dialogBackgroundCaption';
-    const DIALOG_LINK_CAPTION = 'dxHtmlEditor-dialogLinkCaption';
+
     const DIALOG_LINK_FIELD_URL = 'dxHtmlEditor-dialogLinkUrlField';
     const DIALOG_LINK_FIELD_TEXT = 'dxHtmlEditor-dialogLinkTextField';
     const DIALOG_LINK_FIELD_TARGET = 'dxHtmlEditor-dialogLinkTargetField';
     const DIALOG_LINK_FIELD_TARGET_CLASS = 'dx-formdialog-field-target';
-    const DIALOG_IMAGE_CAPTION = 'dxHtmlEditor-dialogImageCaption';
+
     const DIALOG_IMAGE_FIELD_URL = 'dxHtmlEditor-dialogImageUrlField';
     const DIALOG_IMAGE_FIELD_ALT = 'dxHtmlEditor-dialogImageAltField';
     const DIALOG_IMAGE_FIELD_WIDTH = 'dxHtmlEditor-dialogImageWidthField';
     const DIALOG_IMAGE_FIELD_HEIGHT = 'dxHtmlEditor-dialogImageHeightField';
     const DIALOG_TABLE_FIELD_COLUMNS = 'dxHtmlEditor-dialogInsertTableRowsField';
     const DIALOG_TABLE_FIELD_ROWS = 'dxHtmlEditor-dialogInsertTableColumnsField';
-    const DIALOG_TABLE_CAPTION = 'dxHtmlEditor-dialogInsertTableCaption';
+
 
     const USER_ACTION = 'user';
-    const SILENT_ACTION = 'silent';
+
 
     const ICON_MAP = {
         insertHeaderRow: 'header',
@@ -74,17 +71,12 @@ if(Quill) {
         return localize(value) || value;
     };
 
-    const getTargetTableNode = (quill, partName) => {
-        const [table, cell] = quill.getModule('table').getTable();
-        return partName === 'table' ? table.domNode : cell.domNode;
-    };
-
     ToolbarModule = class ToolbarModule extends BaseModule {
         constructor(quill, options) {
             super(quill, options);
 
             this._toolbarWidgets = new WidgetCollector();
-            this._formatHandlers = this._getFormatHandlers();
+            this._formatHandlers = getFormatHandlers.bind(this)();
             this._tableFormats = getTableFormats(quill);
 
             if(isDefined(options.items)) {
@@ -146,115 +138,6 @@ if(Quill) {
             this._toggleClearFormatting(isApplied || !isEmptyObject(formats));
         }
 
-        _getFormatHandlers() {
-            return {
-                clear: ({ event }) => {
-                    const range = this.quill.getSelection();
-                    if(range) {
-                        this.saveValueChangeEvent(event);
-                        this.quill.removeFormat(range);
-                        this.updateFormatWidgets();
-                    }
-                },
-                link: this._prepareLinkHandler(),
-                image: this._prepareImageHandler(),
-                color: this._prepareColorClickHandler('color'),
-                background: this._prepareColorClickHandler('background'),
-                orderedList: this._prepareShortcutHandler('list', 'ordered'),
-                bulletList: this._prepareShortcutHandler('list', 'bullet'),
-                alignLeft: this._prepareShortcutHandler('align', 'left'),
-                alignCenter: this._prepareShortcutHandler('align', 'center'),
-                alignRight: this._prepareShortcutHandler('align', 'right'),
-                alignJustify: this._prepareShortcutHandler('align', 'justify'),
-                codeBlock: this._getDefaultClickHandler('code-block'),
-                undo: ({ event }) => {
-                    this.saveValueChangeEvent(event);
-                    this.quill.history.undo();
-                },
-                redo: ({ event }) => {
-                    this.saveValueChangeEvent(event);
-                    this.quill.history.redo();
-                },
-                increaseIndent: ({ event }) => {
-                    this._applyFormat(['indent', '+1', USER_ACTION], event);
-                },
-                decreaseIndent: ({ event }) => {
-                    this._applyFormat(['indent', '-1', USER_ACTION], event);
-                },
-                superscript: this._prepareShortcutHandler('script', 'super'),
-                subscript: this._prepareShortcutHandler('script', 'sub'),
-                insertTable: this._prepareInsertTableHandler(),
-                insertHeaderRow: getTableOperationHandler(this.quill, 'insertHeaderRow'),
-                insertRowAbove: getTableOperationHandler(this.quill, 'insertRowAbove'),
-                insertRowBelow: getTableOperationHandler(this.quill, 'insertRowBelow'),
-                insertColumnLeft: getTableOperationHandler(this.quill, 'insertColumnLeft'),
-                insertColumnRight: getTableOperationHandler(this.quill, 'insertColumnRight'),
-                deleteColumn: getTableOperationHandler(this.quill, 'deleteColumn'),
-                deleteRow: getTableOperationHandler(this.quill, 'deleteRow'),
-                deleteTable: getTableOperationHandler(this.quill, 'deleteTable'),
-                cellProperties: () => {
-                    const domNode = getTargetTableNode(this.quill, 'cell');
-                    showCellPropertiesForm(this.editorInstance, $(domNode));
-                },
-                tableProperties: () => {
-                    const domNode = getTargetTableNode(this.quill, 'table');
-                    showTablePropertiesForm(this.editorInstance, $(domNode));
-                }
-            };
-        }
-
-        _prepareShortcutHandler(name, shortcutValue) {
-            return ({ event }) => {
-                const formats = this.quill.getFormat();
-                const value = formats[name] === shortcutValue ? false : shortcutValue;
-
-                this._applyFormat([name, value, USER_ACTION], event);
-                this.updateFormatWidgets(true);
-            };
-        }
-
-        _prepareLinkHandler() {
-            return () => {
-                this.quill.focus();
-
-                const selection = this.quill.getSelection();
-                const hasEmbedContent = this._hasEmbedContent(selection);
-                const formats = selection ? this.quill.getFormat() : {};
-                const formData = {
-                    href: formats.link || '',
-                    text: selection && !hasEmbedContent ? this.quill.getText(selection) : '',
-                    target: Object.prototype.hasOwnProperty.call(formats, 'target') ? !!formats.target : true
-                };
-                this.editorInstance.formDialogOption('title', localizationMessage.format(DIALOG_LINK_CAPTION));
-
-                const promise = this.editorInstance.showFormDialog({
-                    formData: formData,
-                    items: this._getLinkFormItems(selection)
-                });
-
-                promise.done((formData, event) => {
-                    if(selection && !hasEmbedContent) {
-                        const text = formData.text || formData.href;
-                        const { index, length } = selection;
-
-                        formData.text = undefined;
-                        this.saveValueChangeEvent(event);
-
-                        length && this.quill.deleteText(index, length, SILENT_ACTION);
-                        this.quill.insertText(index, text, 'link', formData, USER_ACTION);
-                        this.quill.setSelection(index + text.length, 0, USER_ACTION);
-                    } else {
-                        formData.text = !selection && !formData.text ? formData.href : formData.text;
-                        this._applyFormat(['link', formData, USER_ACTION], event);
-                    }
-                });
-
-                promise.fail(() => {
-                    this.quill.focus();
-                });
-            };
-        }
-
         _hasEmbedContent(selection) {
             return !!selection && this.quill.getText(selection).trim().length < selection.length;
         }
@@ -279,52 +162,6 @@ if(Quill) {
             ];
         }
 
-        _prepareImageHandler() {
-            return () => {
-                const formData = this.quill.getFormat();
-                const isUpdateDialog = Object.prototype.hasOwnProperty.call(formData, 'imageSrc');
-                const defaultIndex = this._defaultPasteIndex;
-
-                if(isUpdateDialog) {
-                    const { imageSrc } = this.quill.getFormat(defaultIndex - 1, 1);
-
-                    formData.src = formData.imageSrc;
-                    delete formData.imageSrc;
-
-                    if(!imageSrc || defaultIndex === 0) {
-                        this.quill.setSelection(defaultIndex + 1, 0, SILENT_ACTION);
-                    }
-                }
-
-                const formatIndex = this._embedFormatIndex;
-
-                this.editorInstance.formDialogOption('title', localizationMessage.format(DIALOG_IMAGE_CAPTION));
-
-                const promise = this.editorInstance.showFormDialog({
-                    formData: formData,
-                    items: this._imageFormItems
-                });
-
-                promise
-                    .done((formData, event) => {
-                        let index = defaultIndex;
-
-                        this.saveValueChangeEvent(event);
-
-                        if(isUpdateDialog) {
-                            index = formatIndex;
-                            this.quill.deleteText(index, 1, SILENT_ACTION);
-                        }
-
-                        this.quill.insertEmbed(index, 'extendedImage', formData, USER_ACTION);
-                        this.quill.setSelection(index + 1, 0, USER_ACTION);
-                    })
-                    .always(() => {
-                        this.quill.focus();
-                    });
-            };
-        }
-
         get _insertTableFormItems() {
             return [
                 {
@@ -344,44 +181,6 @@ if(Quill) {
                     label: { text: localizationMessage.format(DIALOG_TABLE_FIELD_ROWS) }
                 }
             ];
-        }
-
-        _prepareInsertTableHandler() {
-            return () => {
-                const formats = this.quill.getFormat();
-                const isTableFocused = this._tableFormats.some(
-                    format => Object.prototype.hasOwnProperty.call(formats, format)
-                );
-                const formData = { rows: 1, columns: 1 };
-
-                if(isTableFocused) {
-                    this.quill.focus();
-                    return;
-                }
-
-                this.editorInstance.formDialogOption('title', localizationMessage.format(DIALOG_TABLE_CAPTION));
-
-                const promise = this.editorInstance.showFormDialog({
-                    formData,
-                    items: this._insertTableFormItems
-                });
-
-                promise
-                    .done((formData, event) => {
-                        this.quill.focus();
-
-                        const table = this.quill.getModule('table');
-                        if(table) {
-                            this.saveValueChangeEvent(event);
-
-                            const { columns, rows } = formData;
-                            table.insertTable(columns, rows);
-                        }
-                    })
-                    .always(() => {
-                        this.quill.focus();
-                    });
-            };
         }
 
         get _embedFormatIndex() {
@@ -575,32 +374,6 @@ if(Quill) {
             if(this.toolbarInstance.option('overflowMenuVisible')) {
                 this.toolbarInstance.option('overflowMenuVisible', false);
             }
-        }
-
-        _prepareColorClickHandler(name) {
-            return () => {
-                const formData = this.quill.getFormat();
-                const caption = name === 'color' ? DIALOG_COLOR_CAPTION : DIALOG_BACKGROUND_CAPTION;
-                this.editorInstance.formDialogOption('title', localizationMessage.format(caption));
-                const promise = this.editorInstance.showFormDialog({
-                    formData: formData,
-                    items: [{
-                        dataField: name,
-                        editorType: 'dxColorView',
-                        editorOptions: {
-                            focusStateEnabled: false
-                        },
-                        label: { visible: false }
-                    }]
-                });
-
-                promise.done((formData, event) => {
-                    this._applyFormat([name, formData[name], USER_ACTION], event);
-                });
-                promise.fail(() => {
-                    this.quill.focus();
-                });
-            };
         }
 
         _getToolbarItem(item) {
